+++ db/db_iter.cc
@@ -117,6 +117,7 @@ class DBIter: public Iterator {
         current_entry_is_merged_(false),
         statistics_(cf_options.statistics),
         version_number_(version_number),
+        iterate_lower_bound_(read_options.iterate_lower_bound),
         iterate_upper_bound_(read_options.iterate_upper_bound),
         prefix_same_as_start_(read_options.prefix_same_as_start),
         pin_thru_lifetime_(read_options.pin_data),
@@ -275,6 +276,7 @@ class DBIter: public Iterator {
   uint64_t max_skippable_internal_keys_;
   uint64_t num_internal_keys_skipped_;
   uint64_t version_number_;
+  const Slice* iterate_lower_bound_;
   const Slice* iterate_upper_bound_;
   IterKey prefix_start_buf_;
   Slice prefix_start_key_;
@@ -686,6 +688,14 @@ void DBIter::PrevInternal() {
       return;
     }
 
+    if (iterate_lower_bound_ != nullptr &&
+        user_comparator_->Compare(saved_key_.GetUserKey(),
+                                  *iterate_lower_bound_) < 0) {
+      // We've iterated earlier than the user-specified lower bound.
+      valid_ = false;
+      return;
+    }
+
     if (FindValueForCurrentKey()) {
       valid_ = true;
       if (!iter_->Valid()) {
@@ -1077,6 +1087,10 @@ void DBIter::SeekToFirst() {
   // because prefix seek will be used.
   if (prefix_extractor_ != nullptr) {
     max_skip_ = std::numeric_limits<uint64_t>::max();
+  }
+  if (iterate_lower_bound_ != nullptr) {
+    Seek(*iterate_lower_bound_);
+    return;
   }
   direction_ = kForward;
   ReleaseTempPinnedData();